[ARCTIC-585][Hive][Flink][Spark]FIX: Adapt hive column name with "$" … (#590)
[ARCTIC-585][Hive][Flink][Spark]FIX: Adapt hive column name with "$" (#588)

* Adapt hive column name with "$"

* Adapt hive column name with "$"

(cherry picked from commit df20033f28800114d43ca3518fab934f4525f095)
