+++ hive/src/main/java/org/apache/iceberg/parquet/AdaptHiveTypeToMessageType.java
@@ -36,6 +36,11 @@ import static org.apache.parquet.schema.PrimitiveType.PrimitiveTypeName.INT32;
 import static org.apache.parquet.schema.PrimitiveType.PrimitiveTypeName.INT64;
 import static org.apache.parquet.schema.PrimitiveType.PrimitiveTypeName.INT96;
 
+/**
+ * Copy from iceberg {@link org.apache.iceberg.parquet.TypeToMessageType} for follow reason:
+ * 1. Make timestamp to int96.
+ * 2. remove {@link AvroSchemaUtil#makeCompatibleName(String)} of name.
+ */
 public class AdaptHiveTypeToMessageType {
   public static final int DECIMAL_INT32_MAX_DIGITS = 9;
   public static final int DECIMAL_INT64_MAX_DIGITS = 18;
@@ -55,7 +60,7 @@ public class AdaptHiveTypeToMessageType {
       builder.addField(field(field));
     }
 
-    return builder.named(AvroSchemaUtil.makeCompatibleName(name));
+    return builder.named(name);
   }
 
   public GroupType struct(org.apache.iceberg.types.Types.StructType struct,
@@ -66,7 +71,7 @@ public class AdaptHiveTypeToMessageType {
       builder.addField(field(field));
     }
 
-    return builder.id(id).named(AvroSchemaUtil.makeCompatibleName(name));
+    return builder.id(id).named(name);
   }
 
   public Type field(org.apache.iceberg.types.Types.NestedField field) {
@@ -96,7 +101,7 @@ public class AdaptHiveTypeToMessageType {
     return Types.list(repetition)
         .element(field(elementField))
         .id(id)
-        .named(AvroSchemaUtil.makeCompatibleName(name));
+        .named(name);
   }
 
   public GroupType map(org.apache.iceberg.types.Types.MapType map, Type.Repetition repetition, int id, String name) {
@@ -106,7 +111,7 @@ public class AdaptHiveTypeToMessageType {
         .key(field(keyField))
         .value(field(valueField))
         .id(id)
-        .named(AvroSchemaUtil.makeCompatibleName(name));
+        .named(name);
   }
 
   public Type primitive(org.apache.iceberg.types.Type.PrimitiveType primitive,
