+++ style/suppressions.xml
@@ -28,5 +28,7 @@
     <suppress checks="naming" files="ArcticIndexHBaseWriter"/>
     <suppress checks="naming" files="ArcticRowDataHBaseLookupFunction"/>
     <suppress checks="naming" files="ShuffleRulePolicy"/>
+    <suppress checks="naming" files="AdaptHiveParquetWithFlinkSchemaVisitor"/>
+    <suppress checks="naming" files="AdaptHiveParquetWithSparkSchemaVisitor"/>
 </suppressions>
